src/main/java/it/lucabaggi/shakespeareanpokemon/client/RestTemplateErrorHandler.java
package it.lucabaggi.shakespeareanpokemon.client;

import org.springframework.http.HttpStatus;
import org.springframework.http.client.ClientHttpResponse;
import org.springframework.stereotype.Component;
import org.springframework.web.client.HttpClientErrorException;
import org.springframework.web.client.ResponseErrorHandler;

import java.io.IOException;

@Component
public class RestTemplateErrorHandler implements ResponseErrorHandler {

    @Override
    public boolean hasError(ClientHttpResponse clientHttpResponse) throws IOException {
        return clientHttpResponse.getStatusCode().isError();
    }

    @Override
    public void handleError(ClientHttpResponse clientHttpResponse) throws IOException {
        HttpStatus httpStatus = clientHttpResponse.getStatusCode();
        String httpStatusText = clientHttpResponse.getStatusText();
        throw new HttpClientErrorException(httpStatus, httpStatusText);
    }
}
